requirements.txt
streamlit>=1.22.0
pydantic>=2.0.0
plotly>=5.14.0
pandas>=2.0.0
supabase>=1.0.3
unstructured>=0.8.0
unstructured-inference>=0.5.0
langchain>=0.1.0
langchain-community>=0.0.10
openai>=0.27.8
numpy>=1.24.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
aiohttp>=3.8.5
asyncio>=3.4.3
pi-heif>=0.22.0
pdf2image>=1.17.0
python-docx>=1.1.2
tiktoken>=0.9.0
pytest-asyncio==0.23.5 